Added skill for convert the value BTC into trading value (USD/EUR)
I utilized this https://developer.bitaps.com/market APi for getting information on actual trading value btc.

If the api response with the status code 200 the result is this 
```
{
   "total BTC": "0.10000000 BTC",
   "total EUR": "945.91000000 EUR",
   "onchain BTC": "0.10000000 BTC",
   "onchain EUR": "945.91000000 EUR",
   "offchainBTC": "0.00000000 BTC",
   "offchain EUR": "0.00000000 EUR",
   "informations": "The network is testnet, so the EUR not are real :("
}
```

if the status code is different the print is this

```
{
   "total BTC": "0.10000000 BTC",
   "onchain BTC": "0.10000000 BTC",
   "offchainBTC": "0.00000000 BTC",
}
```

So, the plugin work without the network (an example, if the node bitcoin work with satellite blockstream).

This pull request include also this fix
- https://github.com/renepickhardt/c-lightning-plugin-collection/pull/4
- https://github.com/renepickhardt/c-lightning-plugin-collection/pull/3